AutoReply/pyproject.toml
[project]
name = "autoreply"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google>=3.0.0",
    "google-genai>=1.49.0",
    "openai-agents>=0.5.0",
    "pyautogui>=0.9.54",
    "pyperclip>=1.11.0",
]
